tidy up writing files to disk
